requirements.txt
fastapi
fastapi-sessions
uvicorn>=0.27.0
python-multipart
pydantic
pydantic-extra-types
pydantic[email]
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
websockets>=11.0,<12.0
crewai
crewai-tools 
networkx
chromadb
pypdf>=3.17.1
python-docx
aiofiles
python-dotenv>=1.0.0
